requirements.txt
gradio==4.31.5
Pillow==9.5.0
Pillow==10.3.0
Requests==2.32.2
torch==2.3.0
transformers==4.26.1
